+++ Cargo.toml
@@ -68,6 +68,7 @@ members = [
 	"bridges/snowbridge/primitives/merkle-tree",
 	"bridges/snowbridge/primitives/outbound-queue",
 	"bridges/snowbridge/primitives/verification",
+	"bridges/snowbridge/primitives/test-helper",
 	"bridges/snowbridge/runtime/runtime-common",
 	"bridges/snowbridge/runtime/test-common",
 	"cumulus/bin/pov-validator",
@@ -1258,6 +1259,7 @@ snowbridge-ethereum = { path = "bridges/snowbridge/primitives/ethereum", default
 snowbridge-inbound-queue-primitives = { path = "bridges/snowbridge/primitives/inbound-queue", default-features = false }
 snowbridge-merkle-tree = { path = "bridges/snowbridge/primitives/merkle-tree", default-features = false }
 snowbridge-outbound-queue-primitives = { path = "bridges/snowbridge/primitives/outbound-queue", default-features = false }
+snowbridge-test-helper-primitives = { path = "bridges/snowbridge/primitives/test-helper", default-features = false }
 snowbridge-outbound-queue-runtime-api = { path = "bridges/snowbridge/pallets/outbound-queue/runtime-api", default-features = false }
 snowbridge-outbound-queue-runtime-api-v2 = { path = "bridges/snowbridge/pallets/outbound-queue-v2/runtime-api", default-features = false }
 snowbridge-pallet-ethereum-client = { path = "bridges/snowbridge/pallets/ethereum-client", default-features = false }
