+++ bridges/snowbridge/pallets/system-v2/Cargo.toml
@@ -34,6 +34,7 @@ xcm-executor = { workspace = true }
 snowbridge-core = { workspace = true }
 snowbridge-outbound-queue-primitives = { workspace = true }
 snowbridge-pallet-system = { workspace = true }
+snowbridge-test-helper-primitives = { optional = true, workspace = true }
 
 [dev-dependencies]
 hex = { workspace = true, default-features = true }
@@ -42,6 +43,7 @@ pallet-balances = { workspace = true, default-features = true }
 polkadot-primitives = { workspace = true, default-features = true }
 snowbridge-pallet-outbound-queue-v2 = { workspace = true, default-features = true }
 sp-keyring = { workspace = true, default-features = true }
+
 
 [features]
 default = ["std"]
@@ -56,6 +58,7 @@ std = [
     "snowbridge-core/std",
     "snowbridge-outbound-queue-primitives/std",
     "snowbridge-pallet-system/std",
+    "snowbridge-test-helper-primitives?/std",
     "sp-core/std",
     "sp-io/std",
     "sp-runtime/std",
@@ -76,6 +79,7 @@ runtime-benchmarks = [
     "sp-runtime/runtime-benchmarks",
     "xcm-executor/runtime-benchmarks",
     "xcm/runtime-benchmarks",
+    "snowbridge-test-helper-primitives/runtime-benchmarks",
 ]
 try-runtime = [
     "frame-support/try-runtime",
